Move files around to separate generic code
Now the dataclasses specific code matches the attrs code and
is in its own module. The generic code is in it's own generic
module and the __init__ just imports the concrete instances
