--- datargs/meta.py
@@ -1,6 +1,0 @@
-class AbstractClassProperty:
-    """
-    Abstract property that should be overridden by a class-level constant.
-    """
-
-    __isabstractmethod__ = True
