+++ datargs/records/error.py
@@ -1,0 +1,2 @@
+class NotARecordClass(ValueError):
+    pass
